+++ storage/url/url.go
@@ -77,8 +77,7 @@ func WithAllVersions(isAllVersions bool) Option {
 
 // New creates a new URL from given path string.
 func New(s string, opts ...Option) (*URL, error) {
-	// TODO Change strCut to strings.Cut when minimum required Go version is 1.18
-	scheme, rest, isFound := strCut(s, "://")
+	scheme, rest, isFound := strings.Cut(s, "://")
 	if !isFound {
 		url := &URL{
 			Type:   localObject,
@@ -135,17 +134,6 @@ func New(s string, opts ...Option) (*URL, error) {
 		return nil, err
 	}
 	return url, nil
-}
-
-// strCut slices s around the first instance of sep,
-// returning the text before and after sep.
-// The found result reports whether sep appears in s.
-// If sep does not appear in s, cut returns s, "", false.
-func strCut(s string, sep string) (before string, after string, isFound bool) {
-	if i := strings.Index(s, sep); i >= 0 {
-		return s[:i], s[i+len(sep):], true
-	}
-	return s, "", false
 }
 
 // IsRemote reports whether the object is stored on a remote storage system.